lp-1269 : add missing strategies for partners
