+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/service/validator/posttransition/partnership/UpdateTerm.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.service.validator.posttransition;
+package uk.gov.companieshouse.limitedpartnershipsapi.service.validator.posttransition.partnership;
 
 import org.springframework.stereotype.Component;
 import uk.gov.companieshouse.api.model.payment.Cost;
@@ -6,22 +6,22 @@ import uk.gov.companieshouse.api.model.validationstatus.ValidationStatusError;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.common.PartnershipKind;
 import uk.gov.companieshouse.limitedpartnershipsapi.model.partnership.dto.LimitedPartnershipDto;
 import uk.gov.companieshouse.limitedpartnershipsapi.service.validator.ValidationStatus;
+import uk.gov.companieshouse.limitedpartnershipsapi.service.validator.posttransition.PostTransitionStrategy;
 
 import java.util.List;
 
 @Component
-public class UpdateRegisteredOfficeAddress implements PostTransitionStrategy<LimitedPartnershipDto> {
-
+public class UpdateTerm implements PostTransitionStrategy<LimitedPartnershipDto> {
     @Override
     public String getKind() {
-        return PartnershipKind.UPDATE_PARTNERSHIP_REGISTERED_OFFICE_ADDRESS.getDescription();
+        return PartnershipKind.UPDATE_PARTNERSHIP_TERM.getDescription();
     }
 
     @Override
     public void validate(LimitedPartnershipDto limitedPartnershipDto, List<ValidationStatusError> errorsList, ValidationStatus validationStatus) {
-        if (limitedPartnershipDto.getData().getRegisteredOfficeAddress() == null) {
-            errorsList.add(validationStatus.createValidationStatusError("Registered office address is required",
-                    "data.registeredOfficeAddress"));
+        if (limitedPartnershipDto.getData().getTerm() == null) {
+            errorsList.add(validationStatus.createValidationStatusError("Term is required",
+                    "data.term"));
         }
     }
